chore: convert navbar into reusable component
