--- src/components/Navbar.tsx
@@ -1,3 +1,0 @@
-export default function Navbar() {
-  return <div>Navbar</div>;
-}
