+++ src/styles/Dashboard.module.css
@@ -3,64 +3,6 @@
   background-color: #f9f9f9;
 }
 
-/* navbar */
-.navbar {
-  height: 80px;
-  background: #979483;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0 22px;
-  box-shadow: 0 8px 18px rgba(0, 0, 0, 0.22);
-  position: relative;
-  z-index: 20;
-}
-
-.logoImg {
-  height: 46px;
-  width: auto;
-  filter: brightness(0) invert(1);
-}
-
-.navLinks {
-  display: flex;
-  align-items: center;
-  gap: 1.5rem;
-}
-
-.navLink {
-  display: flex;
-  align-items: center;
-  gap: 0.35rem;
-
-  color: white;
-  text-decoration: none;
-
-  font-size: 0.95rem;
-
-  background: none;
-  border: none;
-
-  cursor: pointer;
-
-  transition: opacity 0.15s ease;
-}
-
-.navLink:visited {
-  color: white;
-}
-
-.navLink:hover {
-  opacity: 0.8;
-  text-decoration: none;
-}
-
-.navIcon {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
 /* hero */
 .hero {
   position: relative;
@@ -594,37 +536,6 @@
 
 /* for mobile */
 @media (max-width: 480px) {
-  .navbar {
-    height: auto;
-    padding: 12px 16px;
-
-    flex-direction: column;
-    align-items: flex-start;
-    gap: 0.75rem;
-  }
-
-  .logoImg {
-    height: 36px;
-  }
-
-  .navLinks {
-    width: 100%;
-    gap: 0.85rem;
-    overflow-x: auto;
-    padding-bottom: 2px;
-  }
-
-  .navLink {
-    flex: 0 0 auto;
-    font-size: 0.85rem;
-    white-space: nowrap;
-  }
-
-  .navIcon svg {
-    width: 18px;
-    height: 18px;
-  }
-
   .hero {
     height: 360px;
   }
